+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Make assignment binding deterministic by rejecting ambiguous unifications.
+
 ## 0.1.0
 
 - Initial scaffolding for the Ruby::Rego gem.